Huge refactoring which addresses the following:
i) removal of the READABLE / WRITABLE distinction.  This was done following a lot of reading about best practices and server controlled parameters in RESTful APIs.  For example in the RedHat discussion [1] where people generally agree it is acceptable to either silently ignore "read only" fields, or choose to throw an error.  We would most likely silently ignore them rather than throw an error, but it means the following:
  a) All sub resources (tags, contacts etc) are ignored in any update operations
  b) Read only fields (created, modified, deleted) are ignored in any update operations
  c) Key is ignored in any creation service
I believe now these 3 governing rules are acceptable, as it simplifies the codebase hugely, and most clients would craft their own using JSON.

ii) removal of as much inheretance as sensible by favoring interfaces over inheretance.  The revised schema design flushed out things like "only Organization and Dataset should be able to have identifiers".  This meant that not all entities can be treated the same.  A heavy refactor and a use of fluid utility classes in this commit illustrate how this can be accomplished.  These appear in the MyBatis layer, and in the tests, where previously heavy inheritence rules, now fluid static methods are used (complete encapsulation seemed unnecessary at this stage)

[1] https://www.redhat.com/archives/rest-practices/2010-April/thread.html#00041