Adds identifier support read/write support for nodes through the UI.
Relaxes the content-type checking on HTTP DELETE to handle certain browsers with Angular-JS.  There is no content for a delete, so no need to care about what type the client may or may not provide, thus wildcard it.  AngularJS on Chrome *only* gives application/XML and appears to be a known angular "feature", but our services explicitly checked for JSON, despite there being no content anyway.