We now have nested properties.  This like metasync populate them, and they need to be trimmed or else we blow all validation.
StringTrimming now trims all properties in the registry model object packages (only).  Reflection is kinda fragile, and using dynabeans can find all kinds of weird stuff - really we want to conservatively trim only the objects we want to.  For example without a package test (as included in this commit) it'll trim the ENUM of COUNTRY UNKNOWN("") to be null, which we do not want.  This is why we use only the model object package.